requirements.txt
Autologging==1.3.2
python-json-logger==2.0.4
gspread==5.3.2
SQLAlchemy==1.4.32
psycopg2==2.9.3
pandas==1.3.5
PyYAML